+++ Task 2.1.5/pom.xml
@@ -42,5 +42,4 @@
             <scope>provided</scope>
         </dependency>
     </dependencies>
-
 </project>
