+++ Task 2.1.5/src/main/java/hiber/config/AppConfig.java
@@ -1,7 +1,5 @@
 package hiber.config;
 
-import hiber.model.Car;
-import hiber.model.User;
 import org.hibernate.SessionFactory;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
@@ -24,41 +22,41 @@ import java.util.Properties;
 @ComponentScan(basePackages = "hiber")
 public class AppConfig {
 
-   private final Environment env;
+    private final Environment env;
 
-   public AppConfig(Environment env) {
-      this.env = env;
-   }
+    public AppConfig(Environment env) {
+        this.env = env;
+    }
 
-   @Bean
-   public DataSource getDataSource() {
-      DriverManagerDataSource dataSource = new DriverManagerDataSource();
-      dataSource.setDriverClassName(Objects.requireNonNull(env.getProperty("db.driver")));
-      dataSource.setUrl(env.getProperty("db.url"));
-      dataSource.setUsername(env.getProperty("db.username"));
-      dataSource.setPassword(env.getProperty("db.password"));
-      return dataSource;
-   }
+    @Bean
+    public DataSource getDataSource() {
+        DriverManagerDataSource dataSource = new DriverManagerDataSource();
+        dataSource.setDriverClassName(Objects.requireNonNull(env.getProperty("db.driver")));
+        dataSource.setUrl(env.getProperty("db.url"));
+        dataSource.setUsername(env.getProperty("db.username"));
+        dataSource.setPassword(env.getProperty("db.password"));
+        return dataSource;
+    }
 
-   @Bean
-   public LocalSessionFactoryBean getSessionFactory() {
-      LocalSessionFactoryBean sessionFactory = new LocalSessionFactoryBean();
-      sessionFactory.setDataSource(getDataSource());
-//      sessionFactory.setPackagesToScan("hiber");
-      Properties hibernateProperties = new Properties();
-      hibernateProperties.put("hibernate.dialect", env.getProperty("hibernate.dialect"));
-      hibernateProperties.put("hibernate.show_sql", env.getProperty("hibernate.show_sql"));
-      hibernateProperties.put("hibernate.hbm2ddl.auto", env.getProperty("hibernate.hbm2ddl.auto"));
-      sessionFactory.setHibernateProperties(hibernateProperties);
-      sessionFactory.setAnnotatedClasses(User.class, Car.class);
+    @Bean
+    public LocalSessionFactoryBean getSessionFactory() {
+        LocalSessionFactoryBean sessionFactory = new LocalSessionFactoryBean();
+        sessionFactory.setDataSource(getDataSource());
+        sessionFactory.setPackagesToScan("hiber");
 
-      return sessionFactory;
-   }
+        Properties hibernateProperties = new Properties();
+        hibernateProperties.put("hibernate.dialect", env.getProperty("hibernate.dialect"));
+        hibernateProperties.put("hibernate.show_sql", env.getProperty("hibernate.show_sql"));
+        hibernateProperties.put("hibernate.hbm2ddl.auto", env.getProperty("hibernate.hbm2ddl.auto"));
+        sessionFactory.setHibernateProperties(hibernateProperties);
 
-   @Bean
-   public HibernateTransactionManager getTransactionManager(SessionFactory getSessionFactory) {
-      HibernateTransactionManager transactionManager = new HibernateTransactionManager();
-      transactionManager.setSessionFactory(getSessionFactory);
-      return transactionManager;
-   }
+        return sessionFactory;
+    }
+
+    @Bean
+    public HibernateTransactionManager getTransactionManager(SessionFactory getSessionFactory) {
+        HibernateTransactionManager transactionManager = new HibernateTransactionManager();
+        transactionManager.setSessionFactory(getSessionFactory);
+        return transactionManager;
+    }
 }
