+++ Task 2.1.5/src/main/java/hiber/dao/UserDao.java
@@ -5,7 +5,10 @@ import hiber.model.User;
 import java.util.List;
 
 public interface UserDao {
-   void add(User user);
-   List<User> listUsers();
-   User findUserBySerialAndModel(int series, String model);
+
+    void add(User user);
+
+    List<User> listUsers();
+
+    User findUserBySerialAndModel(int series, String model);
 }
