+++ Task 2.1.5/src/main/java/hiber/exception/EntityNotFoundException.java
@@ -1,6 +1,7 @@
 package hiber.exception;
 
 public class EntityNotFoundException extends RuntimeException {
+
     public EntityNotFoundException() {
         super("Entity not found");
     }
