+++ Task 2.1.5/src/main/java/hiber/model/Car.java
@@ -1,7 +1,11 @@
 package hiber.model;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 
@@ -9,19 +13,18 @@ import lombok.ToString;
 @Getter
 @Setter
 @ToString
+@NoArgsConstructor
 public class Car {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
     private String model;
     private int series;
-
-    public Car() {}
 
     public Car(String model, int series) {
         this.model = model;
         this.series = series;
     }
-
 }
