+++ Task 2.1.5/src/main/java/hiber/service/UserService.java
@@ -5,7 +5,10 @@ import hiber.model.User;
 import java.util.List;
 
 public interface UserService {
+
     void add(User user);
+
     List<User> listUsers();
+
     User findUserBySerialAndModel(int series, String model);
 }
