--- Task 2.1.5/src/main/java/hiber/service/UserServiceImp.java
@@ -1,34 +1,0 @@
-package hiber.service;
-
-import hiber.dao.UserDao;
-import hiber.model.User;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-
-@Service
-public class UserServiceImp implements UserService {
-
-   @Autowired
-   private UserDao userDao;
-
-   @Transactional
-   @Override
-   public void add(User user) {
-      userDao.add(user);
-   }
-
-   @Transactional(readOnly = true)
-   @Override
-   public List<User> listUsers() {
-      return userDao.listUsers();
-   }
-
-   @Transactional
-   @Override
-   public User findUserBySerialAndModel(int series, String model) {
-      return userDao.findUserBySerialAndModel(series, model);
-   }
-}
